+++ src/App.tsx
@@ -11,7 +11,7 @@ import { Routes } from 'Routes'
 import 'utils/loadIcons'
 import { useDebugHotkey } from 'hooks/useDebug'
 import { AceStyle } from 'debug/commander'
-import { ErrorBoundary } from 'debug/ErrorBoundary'
+import { ErrorBoundary } from 'components/organisms/ErrorBoundary'
 import { useMod } from 'hooks/useMod'
 import { observer } from 'mobx-react'
 import { toJS } from 'mobx'
