+++ src/debug/index.ts
@@ -1,5 +1,3 @@
 export * from './debug.service'
-export * from './object-renderer'
 export * from './delog.service'
-export * from './delog'
 export * from './error'
