+++ src/extra/react.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import inject from 'flinject'
-import { ErrorBoundary } from 'debug/ErrorBoundary'
+import { ErrorBoundary } from 'components/organisms/ErrorBoundary'
 
 React.createElement = inject(React.createElement, ({ target }, type, props, ...children) => {
   if (typeof type !== 'function') return
