+++ src/screens/Main/DebugSubscreens/Components.tsx
@@ -8,7 +8,7 @@ import { Input } from 'components/atoms/Input'
 import { ColoredTag } from 'components/atoms/ColoredTag'
 import { linearGradient } from 'polished'
 import { Select } from 'components/molecules/Select'
-import { ObjectRenderer } from 'debug'
+import { ObjectRenderer } from 'components/organisms/ObjectRenderer'
 import { CommandPrompt } from 'debug/commander/CommandPrompt'
 import { SquareGroupToggle } from 'components/molecules/SquareGroupToggle'
 import { ProviderIcon } from 'core/providers'
